Move c10 mocks to a separate header file.
