+++ csrc/multidevice/communicator.h
@@ -17,39 +17,10 @@
 #include <torch/csrc/distributed/c10d/Backend.hpp>
 #include <torch/csrc/distributed/c10d/TCPStore.hpp>
 #include <torch/csrc/distributed/c10d/Work.hpp>
+#else
+#include <multidevice/c10d_mock.h>
 #endif
 #include <visibility.h>
-
-#if !defined(NVFUSER_DISTRIBUTED)
-namespace c10d {
-class Work : public torch::CustomClassHolder {
- public:
-  void wait() {}
-};
-
-class Backend : public torch::CustomClassHolder {
- public:
-  c10::intrusive_ptr<c10d::Work> barrier() {
-    return c10::make_intrusive<c10d::Work>();
-  }
-  c10::intrusive_ptr<c10d::Work> send(
-      std::vector<at::Tensor>& tensors,
-      int dstRank,
-      int tag) {
-    return c10::make_intrusive<c10d::Work>();
-  }
-  c10::intrusive_ptr<Work> recv(
-      std::vector<at::Tensor>& tensors,
-      int srcRank,
-      int tag) {
-    return c10::make_intrusive<c10d::Work>();
-  }
-};
-
-class TCPStore : public torch::CustomClassHolder {};
-
-} // namespace c10d
-#endif
 
 namespace nvfuser {
